linux: Update containers.config with security labels
Docker containers rely on the security.capability security flag available in
ext2/3/4 filesystems. This is checked when pulling images from the Docker
registry. When the FS is tmpfs, these attributes are available by default
in the extended set, but for the previously mentioned disk filesystems, they
need to be enabled via the kernel config.

Enable the security labels for ext2/3/4.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>