arm: Add preferred docker version
We don't claim Docker support for LS1021. Still, it works.

So this is the first step towards enabling it in SDK 2.0. Without this, it
compiles the default version as delivered by upstream meta-virtualization - 1.5.
This process fails straight from building go-cross version 1.3 instead of
go-cross v1.5. This error is fixed by this patch.

If a user wants to build docker using "bitbake docker", it will do so with the
supported version (instead the one delivered by upstream meta-virtualization
which is currently broken). I am not including the docker recipe in the
fsl-image-virt build target, since there's currently no documented use case.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>